feat(tests): add ExtendedTestsPart1 suite with MightyUI CSS fixtures
Comprehensive test suite validating parser against real-world CSS:
- 41 tests across Selectors, Values, Properties, AtRules
- Fixture-based testing with CSS fragments from MightyUI

Parser fixes for full typed value support:
- ITypedPropertyValue implementation for all value converters
- TypedFunctionValueConverter for rgba/hsl color functions
- GradientFunctionValue for linear-gradient/radial-gradient
- VarValue/CalcValue detection in AnyValueConverter
- TypedAnyValue.CssText preserves original format (not normalized)

Selector fixes:
- NotSelector serialization includes colon prefix
- Vendor pseudo-elements whitelist (38 entries: -webkit-*, -moz-*, -ms-*)
- ::after test corrected to expect CompoundSelector per CSS Selectors Level 4

Test infrastructure:
- tolerateInvalidValues parameter for strict mode testing
- Edge-case tests explicitly use strict mode

+++ src/ExCSS/Tokens/NumberToken.cs
@@ -34,6 +34,6 @@ namespace ExCSS
             }
         }
 
-        public float Value => float.Parse(Data, CultureInfo.InvariantCulture);
+        public double Value => double.Parse(Data, CultureInfo.InvariantCulture);
     }
 }
