+++ src/ExCSS/Values/Number.cs
@@ -9,27 +9,27 @@ namespace ExCSS
         /// <summary>
         ///     Gets a zero value.
         /// </summary>
-        public static readonly Number Zero = new(0f, Unit.Integer);
+        public static readonly Number Zero = new(0d, Unit.Integer);
 
         /// <summary>
         ///     Gets the positive infinite value.
         /// </summary>
-        public static readonly Number Infinite = new(float.PositiveInfinity, Unit.Float);
+        public static readonly Number Infinite = new(double.PositiveInfinity, Unit.Float);
 
         /// <summary>
         ///     Gets the neutral element.
         /// </summary>
-        public static readonly Number One = new(1f, Unit.Integer);
+        public static readonly Number One = new(1d, Unit.Integer);
 
         private readonly Unit _unit;
 
-        public Number(float value, Unit unit)
+        public Number(double value, Unit unit)
         {
             Value = value;
             _unit = unit;
         }
 
-        public float Value { get; }
+        public double Value { get; }
         public bool IsInteger => _unit == Unit.Integer;
 
         public string CssText => ToString();
